+++ CHANGELOG.md
@@ -26,6 +26,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Send notification to user in case of error
 - Clear user's state when error occurs
 
+### Updated
+- Directory paths
+
 ## [0.1.0] - 2025-05-21 - Vlad
 ### Updated
 - Separate logging for bot logic, aiogram, API. Bot's console logging includes errors from aiogram